Refactor use cases
