+++ src/main/java/com/emazon/ms_stock/domain/use_cases/BrandUseCase.java
@@ -1,7 +1,8 @@
 package com.emazon.ms_stock.domain.use_cases;
 
-import com.emazon.ms_stock.application.dto.PageDTO;
-import com.emazon.ms_stock.application.dto.PageHandler;
+import com.emazon.ms_stock.application.dto.handlers.PageDTO;
+import com.emazon.ms_stock.application.dto.handlers.PageHandler;
+import com.emazon.ms_stock.application.utils.ParsingUtils;
 import com.emazon.ms_stock.domain.api.IBrandServicePort;
 import com.emazon.ms_stock.domain.model.Brand;
 import com.emazon.ms_stock.domain.spi.IBrandPersistencePort;
@@ -19,15 +20,19 @@ public class BrandUseCase implements IBrandServicePort {
 
     @Override
     public void save(Brand brand) {
-        Optional<Brand> opt = persistencePort.findByName(brand.getName());
-
-        if (opt.isPresent()) throw new BrandAlreadyExists();
+        validateBrand(brand);
 
         persistencePort.save(brand);
     }
 
+    private void validateBrand(Brand brand) {
+        Optional<Brand> opt = persistencePort.findByName(brand.getName());
+
+        if (opt.isPresent()) throw new BrandAlreadyExists();
+    }
+
     @Override
     public PageDTO<Brand> findAllPageable(PageHandler page) {
-        return persistencePort.findAllPageable(page);
+        return persistencePort.findAllPageable(ParsingUtils.toPageable(page));
     }
 }
