+++ src/main/java/com/emazon/ms_stock/domain/use_cases/CategoryUseCase.java
@@ -1,8 +1,9 @@
 package com.emazon.ms_stock.domain.use_cases;
 
 
-import com.emazon.ms_stock.application.dto.PageDTO;
-import com.emazon.ms_stock.application.dto.PageHandler;
+import com.emazon.ms_stock.application.dto.handlers.PageDTO;
+import com.emazon.ms_stock.application.dto.handlers.PageHandler;
+import com.emazon.ms_stock.application.utils.ParsingUtils;
 import com.emazon.ms_stock.domain.api.ICategoryServicePort;
 import com.emazon.ms_stock.domain.model.Category;
 import com.emazon.ms_stock.domain.spi.ICategoryPersistencePort;
@@ -20,17 +21,19 @@ public class CategoryUseCase implements ICategoryServicePort {
 
     @Override
     public void save(Category category) {
-        Optional<Category> optCategory = persistencePort.findByName(category.getName());
-
-        if (optCategory.isPresent()) {
-            throw new CategoryAlreadyExists();
-        }
+        validateCategory(category);
 
         persistencePort.save(category);
     }
 
+    private void validateCategory(Category category) {
+        Optional<Category> optCategory = persistencePort.findByName(category.getName());
+
+        if (optCategory.isPresent()) throw new CategoryAlreadyExists();
+    }
+
     @Override
     public PageDTO<Category> findAllPageable(PageHandler page) {
-        return persistencePort.findAllPageable(page);
+        return persistencePort.findAllPageable(ParsingUtils.toPageable(page));
     }
 }
